requirements.txt
# Price Bot Application Dependencies
#
# Core dependencies for Austrian e-commerce price monitoring application
# - Web scraping: BeautifulSoup, Playwright for browser automation
# - Google Sheets API: gspread and Google auth libraries
# - Network optimization: aiohappyeyeballs for DNS resolution
#
# Install with: pip install -r requirements.txt

beautifulsoup4>=4.12.0
playwright>=1.40.0
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
aiohappyeyeballs>=2.6.1